requirements.txt
fastapi==0.115.14
uvicorn==0.34.3
requests==2.32.3
pydantic==2.9.2
pydantic-settings==2.6.1
anitopy==2.1.1
python-dateutil==2.9.0.post0
